refactor(get-endpoint): cleaned up validator test file structure
